Put example apps under separate packages
