+++ jpro-file/example/build.gradle
@@ -17,8 +17,8 @@ javafx {
     modules = ['javafx.controls', 'javafx.swing', 'javafx.media']
 }
 
-//mainClassName = 'one.jpro.platform.file.example.FilePickerSample'
-mainClassName = 'one.jpro.platform.file.example.FileDropperSample'
+//mainClassName = 'one.jpro.platform.file.example.picker.FilePickerSample'
+mainClassName = 'one.jpro.platform.file.example.dropper.FileDropperSample'
 
 application {
     mainClass = "$mainClassName"
