+++ jpro-file/example/src/main/java/module-info.java
@@ -11,5 +11,6 @@ module one.jpro.platform.file.example {
     requires atlantafx.base;
     requires org.slf4j;
 
-    exports one.jpro.platform.file.example;
+    exports one.jpro.platform.file.example.dropper;
+    exports one.jpro.platform.file.example.picker;
 }
