+++ jpro-file/example/src/main/java/one/jpro/platform/file/example/picker/FileTableView.java
@@ -1,4 +1,4 @@
-package one.jpro.platform.file.example;
+package one.jpro.platform.file.example.picker;
 
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
